feat: Fruit 삭제 및 Juice 생성, choiceJuice case 정리
