+++ JuiceMaker/JuiceMaker/Model/Fruit.swift
@@ -13,5 +13,4 @@ enum Fruit: String {
     case pineapple = "파인애플"
     case kiwi = "키위"
     case mango = "망고"
-    
 }
