+++ JuiceMaker/JuiceMaker/Model/JuiceMaker.swift
@@ -9,33 +9,31 @@ import Foundation
 // 쥬스 메이커 타입
 struct JuiceMaker {
     enum Menu: String {
-           case strawberry = "딸기"
-           case banana = "바나나"
-           case kiwi = "키위"
-       }
-       
-       let fruitStore: FruitStore
-       
-       init(fruitStore: FruitStore) {
-           self.fruitStore = fruitStore
-       }
-       
-       func makeJuice(juice: Menu) throws {
-           let recipe = chooseRecipe(of: juice)
-           
-           fruitStore.checkFruitStock()
-           
-       }
-       
-    private func chooseRecipe(of juice: Menu) -> FruitStore.Juice {
-           switch juice {
-           case .strawberry:
-               return [Menu.strawberry: -16]
+        case strawberry = "딸기"
+        case banana = "바나나"
+        case kiwi = "키위"
+    }
 
-           case .banana:
-               <#code#>
-           case .kiwi:
-               <#code#>
-           }
+    let fruitStore: FruitStore
+
+    init(fruitStore: FruitStore) {
+        self.fruitStore = fruitStore
+    }
+
+    func makeJuice(juice: Menu) throws {
+        let recipe = chooseRecipe(of: juice)
+
+        fruitStore.checkFruitStock()
+    }
+
+    private func chooseRecipe(of juice: Menu) -> Juice {
+        switch juice {
+        case .strawberry:
+            return .strawberry
+        case .banana:
+            return .banana
+        case .kiwi:
+            return .kiwi
+        }
     }
 }
